feat: added file exist just to be sure it hasnt been removed in milliseconds.

+++ wp-content/mu-plugins/NginxRecursiveCacheClear.php
@@ -43,7 +43,9 @@ class NginxRecursiveCacheClear
             $cacheFiles = [];
             exec($command, $cacheFiles);
             foreach ($cacheFiles as $cacheFile) {
-                unlink($cacheFile);
+                if (file_exists($cacheFile)) {
+                    unlink($cacheFile);
+                }
             }
         }
         return $url;
